requirements.txt
# Gmail API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# LLM providers
anthropic>=0.40.0
google-generativeai>=0.8.0
groq>=0.9.0

# Attachment parsing (Phase 6)
pdfminer.six>=20221105
python-docx>=1.1.0
icalendar>=5.0.0

# Networking (for Telegram notifications)
requests>=2.31.0

# Config
pyyaml>=6.0
